implemented comments
